chore: rename directory to lspUtils, since it has utils beyond just the parser
