+++ src/flows/pipes/RawFluxEditor/index.ts
@@ -1,7 +1,7 @@
 import {
   parse,
   format_from_js_file,
-} from 'src/languageSupport/languages/flux/parser'
+} from 'src/languageSupport/languages/flux/lspUtils'
 import {find} from 'src/shared/contexts/query'
 import View from './view'
 import ReadOnly from './readOnly'
