+++ src/variables/utils/hydrateVars.test.ts
@@ -7,7 +7,7 @@ import {
 } from 'src/variables/mocks'
 import {getMockedParse} from 'src/shared/utils/mocks/mockedParse'
 
-jest.mock('src/languageSupport/languages/flux/parser', () => {
+jest.mock('src/languageSupport/languages/flux/lspUtils', () => {
   return {
     parse: jest.fn(getMockedParse(hydrateVarsAdditionalMappings)),
   }
